# ms-tutorias/docs/swagger.yaml
openapi: 3.0.3
info:
  title: MS Tutorías API
  version: 1.0.0
  description: Contrato mínimo del servicio orquestador que solicita tutorías mediante una Saga con validación de usuarios, bloqueo de agenda y publicación de notificación.
servers:
  - url: http://localhost:3000
    description: Entorno local según configuración del servicio
tags:
  - name: Tutorías
    description: Solicitud y confirmación de tutorías
paths:
  /v1/tutorias:
    post:
      tags:
        - Tutorías
      summary: Solicita una tutoría
      description: Requiere JWT Bearer. Solo usuarios con rol `student` pueden solicitar tutorías; el `idEstudiante` efectivo se toma del claim `sub` del token y no del cuerpo. Requiere el header `Idempotency-Key`; si ya existe una tutoría con esa clave, se retorna esa tutoría sin reejecutar la Saga (no se reenvía la notificación ni se vuelve a bloquear agenda).
      operationId: postTutoria
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitudTutoriaRequest'
            examples:
              solicitud:
                value:
                  idTutor: t09876
                  fechaSolicitada: '2025-06-25T15:00:00.000Z'
                  duracionMinutos: 60
                  materia: Cálculo Multivariable
      responses:
        '201':
          description: Tutoría confirmada (o solicitud idempotente que retorna una tutoría PENDIENTE/CONFIRMADA ya existente).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tutoria'
        '400':
          description: >-
            Falta el header `Idempotency-Key`, obligatorio para esta operación; o el cuerpo tiene
            `idTutor`/`materia` vacíos, `duracionMinutos` no es un número positivo, o
            `fechaSolicitada` no es una fecha ISO 8601 válida o no es futura.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token ausente, mal formado, inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: El token es válido, pero el rol no está autorizado para solicitar tutorías.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Estudiante o tutor no encontrado durante la validación contra `ms-usuarios`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >-
            Horario no disponible o conflicto de reserva reportado por `ms-agenda` (cuerpo `ErrorResponse`); o la
            `Idempotency-Key` corresponde a una tutoría que ya quedó `FALLIDA` en un intento anterior -- se
            retorna esa tutoría directamente sin reejecutar la Saga (cuerpo `Tutoria`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/Tutoria'
        '503':
          description: '`ms-usuarios` o `ms-agenda` no disponible temporalmente por timeout, red o Circuit Breaker abierto.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno o fallo no controlado durante la orquestación/compensación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Tutorías
      summary: Lista las tutorías del estudiante autenticado
      description: >-
        Requiere JWT Bearer. Devuelve todas las tutorías (en cualquier estado) del estudiante
        identificado por el claim `sub` del token, más recientes primero. No hay un camino
        equivalente para tutores todavía.
      operationId: listarTutorias
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado de tutorías del estudiante autenticado (puede ser un array vacío).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tutoria'
        '401':
          description: Token ausente, mal formado, inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/tutorias/{id}:
    get:
      tags:
        - Tutorías
      summary: Consulta el estado de una tutoría (S8)
      description: >-
        Requiere JWT Bearer. Permite consultar el estado de una tutoría después de crearla (por
        ejemplo, una que quedó `PENDIENTE` por una respuesta lenta de una dependencia). Solo el
        estudiante dueño de la tutoría puede consultarla; si no existe o pertenece a otro
        estudiante, responde `404` en ambos casos (no se distingue, para no confirmarle a un
        estudiante autenticado que un `idTutoria` ajeno existe).
      operationId: getTutoriaPorId
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tutoría encontrada y perteneciente al estudiante autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tutoria'
        '401':
          description: Token ausente, mal formado, inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No existe una tutoría con ese id, o no pertenece al estudiante autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Tutorías
      summary: Cancela una tutoría CONFIRMADA
      description: >-
        Requiere JWT Bearer. Solo se puede cancelar una tutoría en estado `CONFIRMADA`. Intenta
        liberar el horario en `ms-agenda` de inmediato (con reintentos); si no lo logra, la
        tutoría igual queda `CANCELADA` y la liberación se reintenta en segundo plano (misma
        infraestructura que la compensación de la Saga). Mismo criterio de ownership que el `GET`:
        404 uniforme para "no existe" y "no es tuya".
      operationId: cancelarTutoria
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tutoría cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tutoria'
        '401':
          description: Token ausente, mal formado, inválido o expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No existe una tutoría con ese id, o no pertenece al estudiante autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: La tutoría existe pero no está en estado `CONFIRMADA` (no se puede cancelar).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Identificador de correlación propagado a usuarios, agenda y eventos de tracking.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: Clave de idempotencia obligatoria. Si ya existe una tutoría creada con la misma clave, se retorna esa tutoría directamente sin reejecutar la Saga.
  schemas:
    SolicitudTutoriaRequest:
      type: object
      required:
        - idTutor
        - fechaSolicitada
        - duracionMinutos
        - materia
      properties:
        idTutor:
          type: string
          example: t09876
        fechaSolicitada:
          type: string
          format: date-time
          example: '2025-06-25T15:00:00.000Z'
        duracionMinutos:
          type: integer
          minimum: 1
          example: 60
        materia:
          type: string
          example: Cálculo Multivariable
      additionalProperties: true
    Tutoria:
      type: object
      description: >-
        DTO de respuesta pública construido explícitamente por el controlador (no la fila cruda de
        PostgreSQL): usa camelCase, igual que el body de la solicitud, y omite `motivoFallo` salvo
        cuando `estado` es `FALLIDA`.
      properties:
        idTutoria:
          type: string
          example: 101
        idEstudiante:
          type: string
          example: e12345
        idTutor:
          type: string
          example: t09876
        tutorNombre:
          type: string
          nullable: true
          description: >-
            Nombre del tutor, resuelto contra ms-usuarios al momento de crear la tutoría. `null`
            en filas creadas antes de que este campo existiera.
          example: Dra. Elena Solano
        fecha:
          type: string
          format: date-time
          example: '2025-06-25T15:00:00.000Z'
        materia:
          type: string
          example: Cálculo Multivariable
        estado:
          type: string
          enum:
            - PENDIENTE
            - CONFIRMADA
            - FALLIDA
            - CANCELADA
          example: CONFIRMADA
        motivoFallo:
          type: string
          description: Solo presente cuando `estado` es `FALLIDA`.
          example: 'Horario no disponible'
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              example: 'No se pudo completar la solicitud: Horario no disponible'